LayoutTests/inspector/script-profiler/event-type-Microtask-expected.txt

CONSOLE MESSAGE: line 8: Promise Microtask Fired
Tests for generating a ScriptProfiler.Event with ScriptProfiler.EventType.Microtask type.


== Running test suite: ScriptProfiler.EventType.Microtask
-- Running test case: TrackingSession
ScriptProfiler.trackingStart
ScriptProfiler.trackingUpdate
PASS: Event type should be Microtask.
ScriptProfiler.trackingComplete
PASS: Profiles should exist when complete.
PASS: Should be 1 profile for this session.
